fix up the bugs
