+++ BlazarUI/app/stylus/components/repo-branch-card.styl
@@ -44,10 +44,6 @@ $details-width = 50%
   width $info-width
   height 100%
 
-.repo-branch-card__inner-wrapper
-  display inline-block
-  width 100%
-
 .repo-branch-card__repo
   padding-top 6px
   font-size 16px
@@ -55,6 +51,18 @@ $details-width = 50%
   white-space nowrap
   text-overflow ellipsis
   max-width calc(100% - 10px)
+
+.repo-branch-card__main
+  padding 10px 20px
+  height $initial-height
+  cursor pointer
+  background-color white
+  transition background $bg-transition-duration
+  width calc(100% - 80px)
+  display inline-block
+
+  & > div
+    display inline-block
 
 .repo-branch-card__build-and-status
   width 80px
@@ -122,13 +130,11 @@ $details-width = 50%
   color #425b76
   font-family HelveticaNeue-bold
 
-.repo-branch-card__details
+.repo-branch-card__triggered-by
   width $details-width
   height 100%
   vertical-align middle
   line-height 20px
-
-.repo-branch-card__triggered-by
   font-family HelveticaNeue-light
   color #425b76
   overflow hidden
@@ -152,7 +158,7 @@ $details-width = 50%
   padding-top 17px
   font-family HelveticaNeue-medium
 
-.repo-branch-card__expanded-header
+.repo-branch-card__details-header
   font-family HelveticaNeue-light
   font-size 18px
   color #425b76
@@ -161,13 +167,13 @@ $details-width = 50%
   & .repo-branch-card__build-number
     font-family HelveticaNeue
 
-.repo-branch-card__expanded-author
+.repo-branch-card__author
   font-size 14px
   color #2965a8
   font-family HelveticaNeue
   padding-left 20px
 
-.repo-branch-card__expanded-module-rows
+.repo-branch-card__module-rows
   margin-top 20px
   box-shadow 0 3px 3px 0 rgba(0, 0, 0, 0.2)
